Add files via upload
This paper presents an electronic system for controlling a DC motor by using a TRIAC and the elevation motor using Arduino UNO. The system comprises an Arduino microcontroller, an H-bridge driver circuit, and, of course, the DC motor and the elevation motor. The aim of the project was to control the speed and elevation of the treadmill. The paper evaluates the performance of the speed control system by analyzing parameters such as motor speed response, speed stability, and control accuracy. The experimental results show that the implemented TRIAC control system achieves accurate speed control with good stability and response characteristics.